quiz_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.5.0
graphene-django==3.1.1
graphql-core==3.2.3